# openapi.yml
openapi: 3.0.1
info:
  title: MarioLaps API
  version: "1.0"
  description: The MarioLaps API
servers:
  - url: /
...
paths:
...
  /races:
    post:
      summary: Create a new race in the round
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/RaceInput"
                - type: object
                  required:
                    - roundId
      responses:
        201:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    items:
                      $ref: "#/components/schemas/Race"
                required:
                  - data
...
  /races/{race}/laps:
    patch:
      summary: Update withLap
      parameters:
        - in: path
          name: race
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/RaceLapInput"
                - type: object
                  required:
                    - withLap
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    items:
                      $ref: "#/components/schemas/Race"
                required:
                  - data
  /races/{race}/types:
    patch:
      summary: Update race type
      parameters:
        - in: path
          name: race
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/RaceTypeInput"
                - type: object
                  required:
                    - type
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    items:
                      $ref: "#/components/schemas/Race"
                required:
                  - data
  /races/{race}/positions:
    post:
      summary: Add score to race
      parameters:
        - in: path
          name: race
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/RacePositionsInput"
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    items:
                      $ref: "#/components/schemas/Race"
                required:
                  - data
...
components:
  schemas:
...
    Race:
      type: object
      properties:
        id:
          type: string
          format: uuid
        withLap:
          type: boolean
        roundId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        positions:
          type: array
          items:
            $ref: "#/components/schemas/UserPosition"
        raceType:
          type: object
          nullable: true
          properties:
            data:
              $ref: "#/components/schemas/RaceType"
    RaceType:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
...
    UserPosition:
      type: object
      properties:
        position:
          type: integer
        userId:
          type: string
          format: uuid
        raceId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    RaceInput:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
    RaceLapInput:
      type: object
      properties:
        withLap:
          type: boolean
    RaceTypeInput:
      type: object
      properties:
        type:
          type: string
          nullable: true
    RacePositionsInput:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        position:
          type: integer